feat(dedicated.nasha): POST and GET folders
ref:MANAGER-10898
